Renamed `neo.io.hdf5io_new` to `neo.io.hdf5io`
